requirements.txt
alembic==1.16.5
blinker==1.9.0
boto3==1.39.13
botocore==1.39.13
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3

# Dependências de teste
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
